+++ src/Components/FavouritePage.js
@@ -1,11 +1,7 @@
 import Papers from "./Papers";
 
-import { useContext } from "react";
-import { MainContext } from "./Main";
-
-const FavouritePage = () => {
+const FavouritePage = ( { handleLikeOrSave, favList, savedList }) => {
     const pubArray = [];
-    const { favList, handleLikeOrSave, savedList }= useContext(MainContext);
 
     // loop through each item that has been saved as favourite in firebase, and if its not the placeholder, then add it to pubArray and display to the page
     favList.forEach((favItem)=> {
