+++ src/Components/SearchForm.js
@@ -1,8 +1,6 @@
-import { useState, useEffect, useContext } from 'react';
-import { MainContext } from './Main';
+import { useState, useEffect } from 'react';
 
-const SearchForm = () => {
-    const { getQueryParams, setNewSearch } = useContext(MainContext);
+const SearchForm = ( {getQueryParams, setNewSearch} ) => {
 
     const [ showForm, setShowForm ] = useState(false);
     const [ userInput, setUserInput ] = useState({subject: '', title: '', year: '', name: '', doi: ''});
